Remove a bogus platform_family? check
'centos' is not a platform family

+++ libraries/provider_hab_package.rb
@@ -112,9 +112,9 @@ class Chef
         def platform_target
           if platform_family?('windows')
             'target=x86_64-windows'
-          elsif platform_family?('rhel') && node['platform_version'].to_f < 6.0
+          elsif platform_family?('rhel') && node['platform_version'].to_i < 6
             'target=x86_64-linux-kernel2'
-          elsif platform_family?('suse') && node['platform_version'].to_f < 6.0
+          elsif platform_family?('suse') && node['platform_version'].to_i < 6
             'target=x86_64-linux-kernel2'
           else
             ''
